backend/requirements.txt
appdirs==1.4.0
Django==1.10.5
django-cors-headers==2.0.2
djangorestframework==3.5.4
djangorestframework-jsonapi==2.1.1
inflection==0.3.1
packaging==16.8
pyparsing==2.1.10
six==1.10.0
